refactor: separate config check related functions
License: MIT
